PDF interpreter - more annotation improvements
Bug #701889 "Annotation objects missing or incorrectly interpreted when appearance stream missing"

Apply the /RD array (if present) to the annotation /Rect before using
the Rect to draw the border.

We were using the /C array in drawborder to set the colour before
stroking the border, there are two problems with this; firstly the /C
array is not used to specify a border colour, secondly we need to set
the colour for certain objects irrespective of whether they have a
border.

Remove annotsetcolor from drawborder and apply it to Squiggly, Strikeout,
Underline, Highlight, Line and Ink annotations in order to set the
colour before drawing the annotation. The specification says that this
is used for; the border of a Link  annotation, the title bar of a
pop-up annotation's pop-up and the background of the annotation's icon
when closed, but in fact Acrobat also uses it to set the colour for all
the annotation types listed above.

There may still be more work needing done here, but this matches all the
test cases I currently have.