Update Visual Studio solution to VS2019.
Our belief is that this should work in all versions of VS from
2010 onwards.

By moving to VS2019, we lose support totally for Windows XP prior
to SP3. If this is a problem, either build using the Makefile, or
pull the old VS2005 solution out of git and use that. If customers
find this to be a problem, we can maintain the old VS2005 solution
alongside the new one with minimal effort.

Default builds lose support for Windows XP SP3/Windows Server 2003,
but the new "ReleaseXP", "DebugXP" and "MementoXP" configurations
will target these platforms (assuming the v141 tools have been
installed too).

These "XP" binaries will be build into {,debug,mem}objxp directories
and will produce binaries with an 'XP' suffix: i.e.
{,debug,mem}bin/{gs,gpcl6,gxps,gpdl}win{32,32}{,c}xp.exe.