PDF interpreter - use PDFDocEncoding for fonts in text annotations
Bug #701889 ""

We were often not applying PDFDocEncoding to fonts when we used them to
draw text for annotations. This commit re-encodes the selected font
which will give a better result.

We probably should also check for UTF16-BE text and use the FallBack
CIDFont in that case, but we'll wait for an example before trying to
code that.