pdfwrite - use correct FontFile key
This was found as part of bug #701997, even if we set -dCompatibilityLevel
to 1.1, so that type 1 fonts are not converted to type 2 (1c, CFF) we
would still write out a FontFile3 key, instead of the correct
FontFile key.

Ghostscript doesn't care about this, but Acrobat refuses to parse the
font from the file when it has the wrong FontFile key.

The test we were using was incorrect anyway, we were checking
ResourcesBeforeUsgae and we should have been checking HaveCFF because
if HaveCFF is true we would convert type 1 to CFF, and if it isn't we
wouldn't.