Bug 701880: tiffsep1 threshold_from_order caused dots in full white.
For patterns with > 256 dots, threshold_from_order would put in 0 value
cells which would then always be imaged. Change this device to (finally)
use the gx_ht_construct_threshold used by the fast_ht thresholding code
so that it should match the other devices, such as pbmraw.

Also vertically invert the use of the threshold array to match the dots
of the other devices.

Add missing dependencies for gdevtsep.c in devs.mak