PDF interpreter - increase robustness in face of invalid annotations
Bug #702016 "Invalid annotation causing an error"

The PDF file has an annotation where the creating application has
inserted the appearance stream directly into the annotation dictionary.
This breaks the code which attempts to parse the annotation dictionary
from an indirect reference into a dictionary object.

This commit tries to recover from such egregiously broken situations by
running the 'oforce' in a stopped context. This is complicated firstly
by the fact that we parse the annotation dictionary in two places, once
to check it for the presence of transparency and once to actually draw
it (if we are rendering annotations).

In addition, the code could potentially leave anything on the stack, so
we can't use a mark/cleartomark pair to tidy up. Instead, use a
(hopefully) unique name object, create it on the stack and in the case
of an error, unwind the stack back to the expected name.

Finally, report the error to the user in a more or less meaningnful way.