Squash warnings (dropped from earlier commit).
These warnings were supposed to have been squashed a few
commits ago, but somehow got lost in a rebase. Thus they
were mentioned in the commit message, but didn't make it
in.

Avoid problems with 'errno' being a special thing that
shouldn't be used as a parameter name.

Explicitly cast time_t to long for printing.