Further refinement for Bug 701880. Make tiffsep better match pbmraw
The previous change attempted to fix the y-phase, but this was not a
180 degree rotation, so it caused the primary angle of the pattern to
shift by 90 degrees. A better correspondence is achieved by inverting
the polarity of the thresholds. This is slightly "darker" than pbmraw
but the angle and phase matches pbmraw closely.

Note that the phase using the CAL module does not match the non-CAL
or pbmraw, but the angle of the pattern does match.