PDF interpreter - don't abort on errors in JBIG2 globals
This was due to Julian's work on comparing jbig2dec with Luratech.

The test file 'normal_716.pdf' and the cluster test file
jbig2_null_segments_and_glyphs.pdf throw errors when trying to read
the JBIG2 globals dictionary. Luratech doesn't cache the globals, while
jbig2dec does, and the problem was that jbig2dec was returning an error
at a point where the PDF interpreter wasn't able to ignore it, while
Luratech deferred the error to a later point, at a time when the
interpreter could, and did, ignore it.

This meant that Luratech continued to render any further content on the
page while jbig2dec did not, resulting in diffs.

This commit simply runs the global context extraction in a stopped
context. If it fails we clean up the stack and exit, which (effectively)
defers the error to the same point as the Luratech decoder.