Undo earlier commit.
A pattern is drawn with the graphic
state that is present at the time
it was created.  Anything that
happens between that time and the actual
fill or stroke with the pattern is not
used.  One of the files that pdfwrite creates
has the following set of operations

/CS0 cs /P0 scn
/GS0 gs
187.5 375 2062.5 916.668 re
f*

where /GS0 gs sets the alpha to 0.5
Adobe incorrectly uses this alpha for the
pattern fill.