Squashed commit fill-stroke3 branch
This branch had several contributors including Robin Watts,
Ken Sharp and Ray Johnston.

The intent was to add a "fill and stroke" path device method
to ensure we get proper rendering when doing fill/stroke
methods with overprint and transparency.  In addition,
adding the method made it possible for pdfwrite to match
the original pdf source file if it also had fill/stroke methods.

The interaction with overprint and transparency led to significant
changes in parts of the code related to those sections.  In particular,
overprint was reworked to ensure that the overprint compositor and
the transparency compositor were aware if it was to be using
the overprint drawn_comps for stroke or for fill.  The code was
rewritten so that the overprint compositor actions only occurred when
an actual operation occurred (e.g. fill, stroke, image, text) as opposed
to every time that the overprint graphic state values changed.

When there is a fill-stroke operation with transparency, depending upon the
graphic state values, a non-isolated knockout group may be pushed.  From this,
an issue with knockout groups within knockout groups was revealed, which was
related to the backdrop selection.  The backdrop selection code was simplified.