Fix segfaults in transparent patterns with -Z@
One gc'ed member (buf) of the structure for patterns with transparency was not
being initialised after allocation, meaning a debug build (which adds a
validation step to the garbager) would attempt to follow an uninitialised
pointer.

Adding the relevant initialisation meant almost identical code in two places,
so use the pre-existing function insteading of duplicating the code.

Segfaults spotted in weekly debug-Z@ test.