Bug 702052: MaxSpots > 58 caused exception with tiffsep device.
The MaxSpots devn parameter was allowed to be set according the the
capability of the tiffsep device (up to 60), but an array to hold the
equivalent CMYK parameter information was not that large. Get rid of
the MAX_DEVICE_PROCESS_COLORS #define and use the device num_std_colorants
and color_info.max_components to determine the maximum number of spot
colorants possible for this and PageSpotColors and update the Devices.htm
document to reflect the actual limit.

Also to reduce potential mischief, add range check error limits on the
DeviceN params based on the device component support, not the absolute max.