Fix bug 702062: Opacity being applied twice.
If an image has an SMask, the alpha was applied twice, during painting of
the transparency group and again when the transparencygroup was popped.
Thanks to Michael Vrhel for the analysis and suggested cure.

If the device has PreserveSMask (e.g., pdfwrite) the SMask will be processed
by the device and we should not push the group, nor change the opacity.