Bug 702068: Fix 'cracks' appearing around transparently blended regions.
As part of rendering SMasks, we take the bbox for the region, and round
it up to make the buffer in pdf14. If (as is frequently the case) the
SMask contents are an image, the PS rendering of that image can result
in it rounding DOWN (due to "any part of a pixel" vs "pixel centre
covered"). This can result in 'cracks' around the edge of smasked
groups.

The fix, implemented here, is to extend the logic that "gridfits"
images within pattern accumulator to also cover smasks.

To make this work we add a new gx_dev_spec_op to detect whether we are
in an smask definition.