Fix Coverity ID 353647: Possible out of bounds access
The testing where separations were added in pdf14_spot_get_color_comp_index
was wrong since it failed to take into account the process colorants.

Also I found that the allocation of sep_name was not checked for NULL which
could lead to dreferencing location 0 if the allocation failed. Note that
there were other instances of this in gdevdevn.c which will be fixed in a
separate commit.