Bug 702041: Recognize pre-multiplied alpha when decoding with OpenJPEG.
The bug's attached PDF has an annotation where the apperance stream consists of
a JPEG2000 image with a softmask embedded in the image data (SMaskInData is
non-zero). The JPEG2000 bitstream provides all three RGB channels and has both
non-pre-multiplied and pre-multiplied alpha. The last alpha channel defined in
the bitstream happens to be the pre-multiplied alpha. Having more than than a
single alpha channel is not allowed according to PDF 2.0 if SMaskInData is
non-zero. So the JPEG2000 bitstream (and by inference the PDF file) is arguably
broken.

Adobe Acrobat renders the annotation as a mostly transparent image consisting
of a few hand-scribbled lines, but Ghostscript renders the annotation as a
solid black square. So a workaround to match Acrobat is needed even if the PDF
is broken.

OpenJPEG hides that the JPEG2000 bitstream has multiple alpha channels by
discarding all alpha channels except the last one defined, so Ghostscript can
only get at the pre-multipled alpha channel. Previously the OpenJPEG glue layer
therefore didn't recognize any alpha channels at all. The PDF interpreter has
its own JPX cdef header decoder in Resources/Init/pdf_draw.ps and it _did_
recognize that the file contained an alpha channel. This mismatch between the
OpenJPEG glue layer and the PDF interpreter meant that no alpha channel data
was propagated to the transparency compositor. It assumed that the alpha
channel was fully opaque and because the RGB channels are all zero the
annotation was rendered as a solid black square.

Now that pre-multiplied alpha channels are recognized both in the OpenJPEG glue
layer and in the PDF interpreter the pre-multiplied alpha channel data which
contains the hand-scribbling is propagated to the transparency compositor which
therefore can render the annotation similar to Acrobat.