Temporarily restore gs_swapcolors
The stroke+fill branch commit removed the definition of gs_swapcolors,
which meant that the merge wouldn't compile.

It seems likely that we could (should ?) use gs_swapcolors_quick
instead, but its hard to be sure and any diffs would be lost in the
very large number of diffs produced by the merge.

So add back gs_swapcolors, this allows us to build again, and then to
produce a baseline. Next we will test replacing gs_swapcolors with
gs_swapcolors_quick to see if there are any differences.

There are some suspicious differences with this commit, a couple of
files that were subtly wrong before and are differently subtly wrong
now, and some problems with stroke+fill with pdfwrite. I'm hoping
that the pdfwrite problems are due to changes made on the stroke+fill
branch that aren't working here because we don't (yet) use the
stroke+fill primitive.