Replace gs_swapcolors() with gs_swapcolors_quick()
gs_swapcolors() was removed as part of the fill+stroke work. Using
gs_swapcolors_quick() results in different rendering, but checking the
diffs there seem to be as many progressions as regressions, possibly
more, so I've decided to go ahead with this

Need to discuss with Michael implementing the stroke+fill stuff.