PDF interpreter - don't abort stream on halftone errors
Bug #702077 "Rangecheck error @720 dpi in sethalftone with SimpleDot"

The supplied PDF file has a legal but stupid Halftone, it has a type 5
halftone that does not specify a child halftone for each ink, for those
inks that it does specify it uses the same halftone for all of them, and
the child haftone has a /Frequency and /Angle of 1.

Chris suggests this is a Harlequin Rip specific halftone which is
intended not to be used, but to be replaced by something in the rip. Of
course this means that the PDF file is now device-specific because any
conforming PDF consumer which halftones should use the crazy halftone
in the file which will lead to terrible output.

However, poor quality isn't the complaint here, its the fact that we
simply cannot create a halftone tile large enough to produce a 1 lpi
screen at 720 dpi, so we throw an error.

I believe this is a reasonable error, so I'm not going to try and
change it. Instead this commit runs sethalftone in a stopped context
and, if it fails, does not abort the current content stream but instead
emits a warning and carries on.