Fix test of uninitialised value in halftone screens
This is a result of bug #702077 and commit :
http://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=f434497caaf2624caabbbc62f36b745e096e95b5

There's nothing wrong with the commit, but a 64-bit release build on
Linux, using the customer file and 720 dpi, resulted in an infinite
loop. Valgrind suggested this was due to an uninitialised variable.

The problem here was that the code assumed that the first iteration
looking for the best match would succeed. However, because the halftone
has a Frequency of 1 the value of 'W' exceeded the maximum of a short,
which led to use running round the loop without finding a valid halftone.

Because we didn't find a valid halftone we would not set phcp to the
values in 'p' (which is set by gx_compute_cell_values) and so we would
test phcp->C which was uninitialised, and could end up going round the
loop retrying with different values of rt endlessly.

This commit simply short-circuits this by checking to see if we have
at least one valid match (rt is not 1) or that we have set phcp to p
(better is true). If we have done neither of these, then exit
immediately with an error to prevent testing phcp->C.