Fix the fallout from the fill+stroke commit
The fill+stroke code adds a new flag to the graphics state
'is_fill_color' which is essentially undocumented. There is a comment in
gscdevn.c at line 773 which says 'for fill_stroke' and another one in
gsstae.c at line 1458 which says 'used by overprint for fill+stroke'.

This is not the entire story however. Its clear that the overprint
changes incorporated into the fill+stroke commit also make extensive
use of 'is_fill_color' even for operations other than fill+stroke (this
is obvious because this interpreter does not yet use the fill+stroke
primitive).

In addition, the interpreter doesn't set this flag, its set at creation
of the graphics state, and set to true. This is a problem for us as
we have the sense of the color array inverted wrt the 'expected'
usage, ie we set color[0] to be the stroke colour, not the fill colour.

We could fix this by inverting is_fill_color, but I'm concerned that
this would need to be done every time we created a new graphics state
and there might be some path through the code which was missed leading
to difficult to track down rendering problems.

So instead this commit inverts our colour storage; we now use color[0]
as the fill colour and color[1] as the stroke colour.

This causes a number of differences; while a few are regressions
(and probably indicate some fault in the interpreter) the vast majority
are progressions and reverse the odd overprinting problems that
concerned me when I first integrated the fill+stroke work.