Revert "Fix leaks after error in gs_main_init_with_args01() (found by memento)."
This reverts commit f35924926cb35f08be5a12ded4a00eb2f42aed3e.

When memory squeezing gpdl, Julian found leaks in the gpdl
startup of the postscript interpreter when allocations failed
in gs_main_init_with_args01().

The initial fix attempted was to make gs_main_init_with_args01()
closedown the interpreter on such errors. This worked for the
tested cases, but hid problems with non-memory related errors
(such as not being able to open a file specified on the command
line) when called by gs (rather than gpdl).

The change attempted resulted in gs_main_finit() being called
twice as gs (and other users of the gs_main APIs) expect to have
to call gs_main_finit() to shut down after gs_main_init_with_args01()
returns an error code.

The real problem is that gpdl is not doing this in the same way as
other callers.

The gs_main layer has existed for donkeys years, so it's probably
best that we just take it as fixed and don't require callers to
start changing stuff now.

Accordingly, I am reverting the attempted fix here, and will
try a different fix for it that only affects gpdl shortly.