Bug701972 -- fix pdfwrite image downscaling bug
Calculate the bits on the row using the input image width,
not the output image width.

I had to add an extra argument to a few functions to propagate the info
down to where it is needed.

Note obscure case involving calling psdf_setup_image_to_mask_filter()
for a Type 4 image and CompatibilityLevel < 1.3.  This change just
adds the extra input_width arg, which doesn't change whatever it was
doing before.  If there is a bug lurking here, it seems unlikely
anybody cares?