Fix segfault cleaning up after gstate clone failure.
For me this happens with MEMENTO_FAILAT=112447 and for Julian it happens with
MEMENTO_FAILAT=115905.

Basically, the order of events was that we'd allocate a new gstate, copy the
contents of the existing gstate to the new one, then copy the dash pattern
(which could fail) and copy the "client_data" (which could and, in this case,
did fail). When either of those failed, we'd jump to the error handling and
exit with the error.

The problem is that, by jumping to the "fail" label, we skip over all the code
that fixes up the reference counts (and various other things) to account for
the newly created gstate.

This commit reorders things so we allocate the gstate, copy the client data,
copy the dash pattern, *then* assign the contents of the old gstate to the new
one, and carry on. Thus, by the time the gstate contents are copied, the
function cannot fail, and everything is sure to be correctly setup.

In the event that either the dash pattern or the client data fail to allocate,
the new gstate is still filled with NULLs, except for the subset of
"gs_gstate_parts" which are allocated and setup correctly so we can safely
free the graphics state object before returning the error.

Also, use NULL for pointer validation instead of 0.