Bug 702130: Fix JPEG memory failure problems.
The SEGV reported by Memento squeezing (which oddly is at 113600
for me, rather than Julian's higher number) is caused by jpeglib
calling 'mem_term' on something that failed to 'mem_init'.

The fix is to use gs_memory_chunk_unwrap rather than target/release,
and to check whether mem == cmem to see whether that actually did
unwrap.

This runs us into a further problem, where we are freeing the
custom memory descriptor using a gc memory pointer, where it was
allocated using a non gc one.

Rather than just changing the free to use the non gc one, we take
the opportunity to ensure that ALL the blocks are in non gc memory.
JPEGlib will take poorly to its memory being gc'd away, and none
of it needs to subject to gc.

Credit goes largely to Chris for this fix.