Sync deep colour blending with recent changes to 8 bit blending.
Release testing reveals problems with deep color blending and
overprint. This is because changes were made to the 8bit
branches of gxblend.c that weren't fully reflected in the
16bit branches.

Reapply those changes here.

While we're at it, address some other minor code cleanups,
including keeping the ordering of parameters the same between 8
and 16 bit to avoid confusion in future.

The command that showed the inital problem was:

bin/gswin32c.exe -o out.psd -sDEVICE=psdcmyk16 tests_private/comparefiles/Bug695916.pdf

where the image of the Earth did not appear.