Bug 702068 continued: Fix smasked images with masks.
Release testing has shown up another case that needs help.

  gs/bin -sDEVICE=ppmraw -o out.ppm -r300 -dMaxBitmap=1G
    tests_private/pdf/sumatra/1901_-_tiling_inconsistencies.pdf

This shows horizontal white lines in the "Next" "Up" "Previous"
images at the top of the page.

Investigation shows this is due to the images having Masks as well
as SMasks.

The Mask image is run to a 1 bit memory device, which means that
the gxdso to check if we are in an SMask doesn't work. We work
around this by introducing a new flag to gs_pixel_image_common
that we can set to indicate that we are within an smask. We set
this when we set the masked image up (using the gxdso), and check
it when we come to do the gridfitting.