Bug 702128 :  clist pdf14 overprint state
This change ensures that gxdso_overprint_active
gets the proper state information.

When the clist writer is active with the pdf14 device, the writer maintains
the current overprint state (as the writer has no access to the pdf14 member variables)
and must know what to put in the clist.

When the clist writer is not present (i.e. page mode), the pdf14 device
maintains the information.