Fix "blue gradients" seen with psdrgb device.
Since the fill/stroke commit went in, some files have been showing
gradients in the wrong colours. For instance this one, spotted in
release testing:

 gs -sDEVICE=psdrgb -r72 -dMaxBitmap=2000M -o out%d.psd
   ../tests_private/comparefiles/Layout.pdf

Bisecting the fill/stroke branch back, it seems the commit that
caused it is the one that stops psdrgb using 'devn' colors.

This causes gradients to drop to using the color_info.comp_bits and
comp_shift information, which the psd devices don't fill in.

The fix here is to make psd_prn_open call the standard function for
setting up these values.