Bug 702117 Knockout groups containing non-isolated groups
So this ended up being rather complicated.   The first issue was that
we were not setting the backdrop properly for non-isolated groups that
reside in knockout groups.  When that was fixed, it was revealed that we
were not doing the recompositing operation for non-isolated knockout groups.  This meant
that the backdrop was getting applied twice effectively (once in the original
fill and then again when the group was popped).  The only file that showed
this issue was the crazy Altona page with the many overlapping groups.  Customer
file added to provide a clearer case in the regression tests.