Squash some warnings with gs_swapcolors_quick breaking const.
At various places in the code, we call gs_swapcolors_quick where
we only have a const pgs.

Properly, we should copy the pgs to a non-const pgs, and then use
that, but that would be a performance hit.

In all cases, we change the graphics state, then restore it before
exit, so effectively it is const over the entire call of the
enclosing procedure.

Rather than having lots of small hacky casts removing const, we
push the const breakage into gs_swapcolors_quick itself where we
do it in the "approved" way using a union.