Fix gx_default_copy_alpha_hl_color for 16 bit operation.
The code to write the composite data back to buffers to be sent
to copy_alpha was getting the first byte of each 16 bit pair
wrong in the 16bit data case.

Also, when we 'restart' after skipping a zero byte in the 16bit
case we need to skip 2 bytes per pixel, not one.

This can be seen when running:

 gs -r72 -o out%d.psd -sDEVICE=psdcmyk16 -dMaxBitmap=2000M
                  ../tests_private/comparefiles/js.pdf