CID 354291: Add error handling for inability to allocate
Coverity complained about setting dash.pattern = NULL and "Passing "&dash"
to "gstate_copy_dash", which dereferences null "dash.pattern" which is either
wrong, or *very* misleading. We assume that it means that gx_set_dash() *may*
access that pointer.

In truth, this can never happen in these circumstances.

Nevertheless, in gx_set_dash() is we have a NULL mem (allocator) pointer and the
dash pattern pointer is NULL at the point we have to copy into it, then throw
an error.