Fix transparency memory allocation bug on Bug689581.pdf
The device was being setup as a framebuffer because the calculation
hadn't been told it would be using transparency (which requires more memory).
Then later it ran out of memory.

Now for this particular page size and memory configuration, it will
properly decide it needs to do a clist.

params to demonstrate the fixed bug:
-dPDFDEBUG -Z: -sDEVICE=pbmraw -dMaxBitmap=2000m -K2000000 -r300 -o out2-%02d.ppm /home/nad/work/samples/tests_private/comparefiles/Bug689581.pdf