Fix stream issue with shading dict
Need to pass in the shading_dict instead of the stream_dict.  This was
only a problem if the stream_dict was compressed, and had a shading
command in it.  I guess fairly uncommon (except for files produced by
pdfwrite).

I don't know why it suddenly started showing up after fill/stroke?  It
would always have been a problem.