Bug 702155 Alternate tint transform and shadings
Add option to match AR shadings for corner case, which
is not as detailed in the PDF specification.

If the source space is DeviceN or Separation and the
colorants are not supported (i.e. if we are using the alternate tint
transform) the interpolation should occur in the source space to
accommodate non-linear tint transform functions.
e.g. We had a case where the transform function
was an increasing staircase. Including that function in the
gradient smoothness calculation gave us severe quantization. AR on
the other hand is doing the interpolation in device color space
and has a smooth result for that case. So AR is not following the spec. The
bit below solves the issues for Type 4 and Type 5 shadings as
this will avoid interpolations in source space. Type 6 and Type 7 will still
have interpolations in the source space even if pfs->cs_always_linear == true.
So the approach below does not solve those issues. To do that
without changing the shading code, we could make a linear
approximation to the alternate tint transform, which would
ensure smoothness like what AR provides.