Reimplement saturation blendmode.
To test the results of deep color rendering, we ran a psdcmyk vs
psdcmyk16 comparison. This showed differences on page 17 of the
Altona_technical_v20_x4.pdf file in one patch. Simplifying the
file showed that the saturation blend mode was showing a
significant difference in one tile.

Saturation blend mode is defined to not make any changes if we
are writing over a pure grey color (as there is no 'hue' for
it to saturate). You'd expect that the blending function would be
continuous (i.e. that a small peturbation of the background color
should only produce a small peturbation in the output), but this
is NOT the case around pure greys.

The example in the tested file, shows that psdcmyk is called with
7a, 7a, 7a, which therefore leaves the background unchanged. For
psdcmyk16, it's called with 7a01 7a03 7a01, which therefore does
NOT leave the background unchanged. Testing by changing the 8 bit
inputs to 7b 7a 7b (a small peturbation), gives output of 99 64 99
(a large change).

So, actually, the results given seem reasonable in that case.

HOWEVER, in investigating what now seems to be a red herring, I
believe I've spotted a problem in the implementation. The scaling
of the components subtracts y when it should be subtracting minb.

The only way I could actually work that out though, was to completely
reimplement the code, and optimise it. I believe the new code is
slightly clearer (in that it follows the spec more explicitly) and
possibly slightly more performant, so I am committing that rather
than attempting to patch the existing code.