Don't treat missing Resources in Pattern as an error
The spec says that the Resources are required, but most readers are
fine without them.  I noticed Adobe doesn't render it but I couldn't
see any kind of warning/error message.

This is the file that was causing a seg fault because it had a null
pattern in it.  The pattern was null ultimately because of the missing
Resources.

I noticed it was gx_remap_Pattern() in the graphics library that was
setting up an invalid pattern that causes a seg fault later in
gs_fill_stroke().  So I think the fix to test for null patterns is
still needed.

Fixes: tests_private/pdf/sumatra/infinite_pattern_recursion.pdf

Note: I am not sure why this file is called
"infinite_pattern_recursion" because I don't see any evidence that it
does that?