More problems with TTF "notdef" glyphs
The font in this file is questionable according to the spec. The spec says GID 0
in a TTF *must* be the "notdef" glyph which, traditionally, is a hollow
rectangle glyph. This contradicts what Postscript and PDF expect which is that
the .notdef glyph will be non-marking.

The problem is many PDF embedded TTF fonts ignore this, and populate GID 0 with
a "meaningful" glyph rather than an actual notdef.

Previously, the fonts that did this were all ones that used MS Unicode cmap
tables and hence we used that to determine whether to image GID 0 or not in
TTFs.

This file does not use a Unicode cmap, but rather an Apple cmap table. In this
case, the cmap table maps every character code to GID 0, GID 0 being the only
glyph defined in the font.

So, tweak the heuristic so that in the event we have what *looks* like a
notdef, we'll look up the character code in the TTF cmap table, and if
the cmap table maps that code to GID 0, then we'll image the glyph.
Otherwise, we will force it to be non-marking.

(Bug 702167)