PS interpreter some 64-bit integer fixes
Bug #702194 "cvi produces incorrect results"

This was simply caused by a cast through long. However its not enough
merely to change the cast, we must also check for running in CPSI mode
and throw an error if the returned integer cannot be represented in a
32-bit word, for compatibility with test suites.

In addition (see bug #702134) when exp throws an undefinedresult error
delay setting the stack variable until after we have tested the result,
so that we do not end up returning a INF on the stack, as is done in
other code paths which check for NaN and INF.

Thirdly; in div cater for the case of an integer divided by a real and
check the result for NaN and INF (as the other cases do). This now
properly throws an error and does not return NaN or INF on the stack.